Fix bugs

+++ CCI/cci_graphem_sp_CV.m
@@ -36,7 +36,7 @@ function cci_graphem_sp_CV(target_spars, Kcv)
 
 	for k = 1:Kcv
 		% Run GraphEM
-		[Xg{k},Mg{k},Cg{k}] = graphem(Xcv{k},opt);
+		[Xg{k},Mg{k},Cg{k}] = graphem(double(Xcv{k}),opt);
 		Xg_k = Xg{k};
 		SPkfoldtag = ['cci_combined_SPCV_sp' num2str(target_spars*100) '_k' num2str(k) '.mat'];
 		SPkfoldpath = [odir SPkfoldtag];
